requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-markdownx==3.0.1
idna==2.8
Markdown==3.2.1
Pillow==7.1.1
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
